* bld/sparc64:
  test, make fsck.21.ut and fsck.22.ut endian-dependent
  test, make fsck.17.ut and fsck.18.ut endian-dependent
  test, make fsck.19.ut and fsck.20.ut endian-dependent
  test, make fsck.23.ut and fsck.24.ut endian-dependent
  test, import fsck.17.ref.beute and rename fsck.17.ref.ute to fsck.17.ref.leute
  hygiene, avoid shadow of global mux symbol (dso entry function)
  hygiene, avoid possible shadowing in ute-mux.c's find_muxer()
  fix, PEBKAC, amend to earlier commit, use sym instead of shadowed pr_sym
  hygiene, avoid shadowing global symbol pr (public dso entry fun)
  hygiene, avoid shadowing the pr_sym() helper in ute-print.c's find_printer()
  hygiene, don't cast consts away in __tai_offs() of tzraw.c
  hygiene, be clear about the signedness of return values in tzraw helpers
  hygiene, cast %jd objects to intmax_t

+++ src/tzraw.h
@@ -226,7 +226,7 @@ static inline int32_t
 zif_trans(zif_t z, int n)
 {
 /* no bound check! */
-	return zif_ntrans(z) > 0 ? be32toh(z->trs[n]) : INT_MIN;
+	return zif_ntrans(z) > 0UL ? (int32_t)be32toh(z->trs[n]) : INT_MIN;
 }
 
 /**
@@ -243,7 +243,7 @@ static inline uint8_t
 zif_type(zif_t z, int n)
 {
 /* no bound check! */
-	return (uint8_t)(zif_ntrans(z) > 0 ? z->tys[n] : 0);
+	return (uint8_t)(zif_ntrans(z) > 0UL ? z->tys[n] : 0U);
 }
 
 /**
